{EQEmu PR 3738} Fix Swarm Pets Requiring NPC Aggros Flag

+++ zone/mob_ai.cpp
@@ -1131,7 +1131,12 @@ void Mob::AI_Process() {
 			SetTarget(hate_list.GetClosestEntOnHateList(this));
 		} else {
 			if (AI_target_check_timer->Check()) {
-				if (IsNPC() && (!IsPet() || (HasOwner() && GetOwner()->IsNPC())) && !CastToNPC()->WillAggroNPCs()) {
+				if (
+					IsNPC() &&
+					!CastToNPC()->GetSwarmInfo() &&
+					(!IsPet() || (HasOwner() && GetOwner()->IsNPC())) &&
+					!CastToNPC()->WillAggroNPCs()
+				) {
 					WipeHateList(true); // wipe NPCs from hate list to prevent faction war
 				}
